requirements.txt
requests
bs4
google-genai
pymongo
python-dotenv